[feat] 涂鸦画板 done： 画笔粗细应该是float而非int

+++ module_reader/src/main/java/com/gracefulwind/learnarms/reader/widget/doodle/DoodleView.java
@@ -263,7 +263,7 @@ public class DoodleView extends View implements Smartable {
         mPresenter.setPaintSize(paintSize);
     }
 
-    public int getPaintSize(){
+    public float getPaintSize(){
         return mPresenter.getPaintSize();
     }
